Rename room_position module to position for clarify vs RoomPosition pure-javascript object
