+++ src/local/position/game_methods.rs
